refactor: moved json parsing into a new file so it's not hidden away
